+++ libsolidity/ast/ASTForward.h
@@ -100,6 +100,25 @@ class StructuredDocumentation;
 
 class VariableScope;
 
+template <class T>
+struct ASTCompareByID
+{
+	using is_transparent = void;
+
+	bool operator()(T const* _lhs, T const* _rhs) const
+	{
+		return _lhs->id() < _rhs->id();
+	}
+	bool operator()(T const* _lhs, int64_t _rhs) const
+	{
+		return _lhs->id() < _rhs;
+	}
+	bool operator()(int64_t _lhs, T const* _rhs) const
+	{
+		return _lhs < _rhs->id();
+	}
+};
+
 // Used as pointers to AST nodes, to be replaced by more clever pointers, e.g. pointers which do
 // not do reference counting but point to a special memory area that is completely released
 // explicitly.