+++ test/libsolidity/syntaxTests/bytecodeReferences/circular_reference_internal_functions.sol
@@ -1,0 +1,3 @@
+contract C { function foo() internal { new D(); } }
+contract D { function foo() internal { new C(); } }
+// ----