+++ test/libsolidity/syntaxTests/bytecodeReferences/circular_reference_libraries.sol
@@ -1,0 +1,3 @@
+library L1 { function foo() internal { L2.foo(); } }
+library L2 { function foo() internal { L1.foo(); } }
+// ----