+++ test/libsolidity/syntaxTests/metaTypes/codeAccessBase.sol
@@ -1,22 +1,26 @@
 contract Base {
     function f() public pure returns (uint) {}
 }
-contract Test is Base {
+contract Test1 is Base {
     function creation() public pure returns (bytes memory) {
-        return type(Test).creationCode;
+        return type(Test1).creationCode;
     }
+}
+contract Test2 is Base {
     function runtime() public pure returns (bytes memory) {
-        return type(Test).runtimeCode;
+        return type(Test2).runtimeCode;
     }
+}
+contract Test3 is Base {
     function creationBase() public pure returns (bytes memory) {
         return type(Base).creationCode;
     }
+}
+contract Test4 is Base {
     function runtimeBase() public pure returns (bytes memory) {
         return type(Base).runtimeCode;
     }
 }
 // ----
-// TypeError 4224: (165-188): Circular reference for contract code access.
-// TypeError 4224: (271-293): Circular reference for contract code access.
-// TypeError 4224: (381-404): Circular reference for contract code access.
-// TypeError 4224: (491-513): Circular reference for contract code access.
+// TypeError 7813: (166-190): Circular reference to contract bytecode either via "new" or "type(...).creationCode" / "type(...).runtimeCode".
+// TypeError 7813: (300-323): Circular reference to contract bytecode either via "new" or "type(...).creationCode" / "type(...).runtimeCode".