+++ test/libsolidity/syntaxTests/nameAndTypeResolution/236_cyclic_binary_dependency.sol
@@ -2,4 +2,6 @@ contract A { function f() public { new B(); } }
 contract B { function f() public { new C(); } }
 contract C { function f() public { new A(); } }
 // ----
-// TypeError 4579: (131-136): Circular reference for contract creation (cannot create instance of derived or same contract).
+// TypeError 7813: (35-40): Circular reference to contract bytecode either via "new" or "type(...).creationCode" / "type(...).runtimeCode".
+// TypeError 7813: (83-88): Circular reference to contract bytecode either via "new" or "type(...).creationCode" / "type(...).runtimeCode".
+// TypeError 7813: (131-136): Circular reference to contract bytecode either via "new" or "type(...).creationCode" / "type(...).runtimeCode".